consolidate video banner comp into homepage and centralize the video player
