+++ src/pages/blog/video-banner/_VideoCarousel.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 // import VideoSrc from '../../markets/static/video/globe.mp4'
-import VideoPlayer from './_video-player'
+import VideoPlayer from '../_video-player'
 import { CustomCarousel } from './carousel/_custom-carousel'
 import { Flex } from 'components/containers'
 import { Header } from 'components/elements'
@@ -196,7 +196,7 @@ const VideoCarousel = ({ carousel_items }) => {
                     </CustomCarousel>
                 </Flex>
                 {is_mobile && (
-                    <SeeMoreBtnMobile to="/">
+                    <SeeMoreBtnMobile to="/blog/videos">
                         <Header type="paragraph-2" color="white" align="center">
                             See all videos
                         </Header>
